test(stringcalculator) : 실수 계산 구현에 따른 테스트 코드 수정

+++ src/test/kotlin/calculator/model/StringCalculatorTest.kt
@@ -9,6 +9,6 @@ class StringCalculatorTest {
     fun addNumbersTest() {
         val calculator = StringCalculator()
         val num = calculator.addNumbers(listOf("1", "105"))
-        assertEquals(106.toBigInteger(), num)
+        assertEquals(106.toBigDecimal(), num)
     }
 }
